issue: unnecessarily applying duty
Sometimes the current duty is read as a non-round float, making a small difference between what was set and what is read from the device.
We just int() the result here, as that's close enough.
Float is helpful for other situations but this caused the scheduler to unnecessarily apply duty settings every 6 seconds, even though the value had not changed, because the difference was greater than 0.