ui: copy favicon into a folder
This keeps the root direction one, that may be preferred for some systems and I believe is the last-change location to find the favicon for browswer. It is now also in a subfolder because Firefox has a strange bug where the Bookmark and History icons weren't working at all when using localhost. Putting it in a subfolder seems to have solved the issue.