feat(deamon): warmup_duration field for alerts to delay activation
`warmup_duration` field is an f64 representing seconds throughout which the
alert condition must hold before the alert is activated. During this time the
alert may go back to inactive state without ever being activated.

`warmup_duration` defaults to 1 for newly created alerts and to 0 for existing
alerts. This is for backwards compatibility.

Warmup is achieved by a new `AlertState`: `WarmUp(DateTime<Local>)`. `WarmUp`
holds the datetime at which we entered the warmup state. We keep track of this
and activate the alert when it's been at least `warmup_duration` seconds since
we entered `WarmUp`. This means we don't guarantee 100% accuracy in handling
`WarmUp` -> `Active` transitions and even 0 `warmup_duration` will be in
`WarmUp` state for a brief period - though this should be invisible to the user
for the most part so I opted not to handle a special case.

There's another new state that is not directly related to this change: `Error`.
In the previous implementation, the alert was put in `Active` state when an
error occured while processing it. That use case is now replaced with `Error`
which, similar to `Activate`, sends an alert message. This allows to us to treat
them differently during state change handling as well as allows clients to
differentiate alert activations from internal errors.

Setting the state is offloaded to a basic state machine in `Alert::set_state`.
This function handles non-error state changes and assumes the error condition
must've been resolved if its called. Any state can transition into an `Error`
and `Error` can transition to `Inactive` or `WarmUp` depending on the condition
at the moment the error is resolved. Ignoring `Error` below are the possible
state transitions:

- `Inactive` -> `WarmUp`
- `WarmUp` -> `Inactive`
- `WarmUp` -> `Active`
- `Active` -> `Inactive`

With the introduction of the state machine and the `Error` state,
`AlertController::activate_alert` and `AlertController::deactivate_alert` are
replaced with `AlertController::activate_alert_with_error`. This function
behaves similar to `activate_alert` but sets the state to `Error`. Since the
non-error state changes are handled elsewhere, the other two functions are no
longer needed.