feat(ui): use poll rate for warmup duration props
Step size will be equal to the poll rate and max will be five times the poll rate.
We mark the field invalid if the chosen duration is not a multiple of the poll rate.
This is not a hard error but just a warning.

The user can technically break things by setting the value manually to an invalid duration. Because
the step will keep inc/decrementing to another invalid value. I tried to find a way to disable
keyboard input on the field but no such option seem to exist.