fix: remove UI-connection-daemon settings
These settings were created when the UI ran normally from a different
location than the daemon. Since the Tauri implementaion has been
removed, we no longer do that - the daemon is the source for both the UI
and the API - so only for local testing is there a need to change these
settings. Changeing them is still possible, as when the UI cannot
connect, it still asks for different settings. These settings in the UI
are simply confusing for 99% of the situations.