fix: resume issue with rate limiter
There is an issue with the Quanta clock, which is used by default, in
the governer. When resuming from suspend after a period of time, the
rate limiter thinks waaaaay too many requests have been make in a short
period blocks all requests for a long time. Aka, the Quanta clock
doesn't reset on resume like the default Instant/system clock does.