feat: add support for direct access
Using the direct_access kwarg results in the drivers spamming warning
logs. Instead of parsing logs and filtering for every possible log
output, we remove the hwmon information from the driver so that it has
no chance to do anything other than direct access, and produces no extra
logs. (the logs are good for the CLI, but can be spamy for CC)